+++ packages/terra-form-textarea/src/terra-dev-site/doc/example/FillContainer.jsx
@@ -1,13 +1,16 @@
 import React from 'react';
-
+import Field from 'terra-form-field';
 // eslint-disable-next-line import/no-extraneous-dependencies, import/no-unresolved, import/extensions
 import Textarea from 'terra-form-textarea/lib/Textarea';
 
 const FillContainerExample = () => (
-  <Textarea
-    size="full"
-    defaultValue="I am full container width and height, making me non-resizable."
-  />
+  <Field label="Fill Container" htmlFor="fill-container">
+    <Textarea
+      size="full"
+      defaultValue="I am full container width and height, making me non-resizable."
+      id="fill-container"
+    />
+  </Field>
 );
 
 export default FillContainerExample;