+++ packages/terra-form-textarea/CHANGELOG.md
@@ -3,7 +3,11 @@ ChangeLog
 
 Unreleased
 ----------
+
+2.34.0 - (January 2, 2019)
+------------------
 ### Changed
+* Minor dependency version bump
 * Removed use of componentWillMount from examples
 * Updated site examples to be more accessible.